agent_requirements.txt
# PDF Agent Server Dependencies
fastapi>=0.104.1
uvicorn>=0.24.0
pypdfium2==4.30.0
pdfminer.six>=20231228
jsonschema>=4.20.0
requests>=2.31.0
pillow>=10.0.1
pydantic>=2.5.0

# Optional: For enhanced PDF processing
pypdf>=3.17.0

# Development dependencies (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Hugging Face Transformers backend (Gemma 3)
torch>=2.2.0
transformers>=4.43.0
accelerate>=0.30.0
sentencepiece>=0.1.99
safetensors>=0.4.2
# Vision backbones used by some HF VLMs
timm>=0.9.12
# Optional: for Qwen2-VL image utilities
# qwen-vl-utils>=0.0.8
# Optional for quantization (set HF_LOAD_IN_4BIT=1)
# bitsandbytes>=0.43.0
